Das2ServerInfoTest:   Remove failing unit test until it's implemented.
GeneralBioSeq, *CharIterator, NibbleIterator:
  Merge *CharIterator interfaces.

CharSeqGlyph:
  Retab and add override annotations.  Use SearchableCharIterator 
interface.  

FlyPointLinkerGlyph, EfficientLabelledLineGlyph:
  Fix warnings.



git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@3748 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ igb/src/com/affymetrix/igb/glyph/EfficientLabelledLineGlyph.java
@@ -59,10 +59,10 @@ public final class EfficientLabelledLineGlyph extends EfficientLabelledGlyph
         //        fillDraw(view);
         if (label_loc == NORTH) {
           g.fillRect(pixelbox.x, pixelbox.y+(pixelbox.height/2),
-                     pixelbox.width, (int)Math.max(1, pixelbox.height/2));
+                     pixelbox.width,Math.max(1, pixelbox.height / 2));
         } else {
           g.fillRect(pixelbox.x, pixelbox.y,
-                     pixelbox.width, (int)Math.max(1, pixelbox.height/2));
+                     pixelbox.width,Math.max(1, pixelbox.height / 2));
         }
       }
       else {
@@ -116,7 +116,7 @@ public final class EfficientLabelledLineGlyph extends EfficientLabelledGlyph
       if (getChildCount() <= 0) {
         // if no children, draw a box
         g.fillRect(pixelbox.x, pixelbox.y + (pixelbox.height / 2),
-                pixelbox.width, (int) Math.max(1, pixelbox.height / 2));
+                pixelbox.width,Math.max(1, pixelbox.height / 2));
       } else {
         // if there are children, draw a line.
         drawDirectedLine(g, pixelbox.x, pixelbox.y + pixelbox.height / 2, pixelbox.width, this.arrowDirection);
